Base files:     * configure.ac: Now version 20201223.     * configure: Regenerated.     * scripts/buildstandardsource.sh: Added a clarifying       comment.     * scripts/ddbuild.sh: Now builds the sourcefile       opstabcount.c for dwarfdump.
dwarfdump/
    * CMakeLists.txt: Adds opscounttab.c and opscounttab.h
      to the files in the build of dwarfdump.
    * Makefile.am: Adds opscounttab.c opscounttab.h and
      buildopscounttab.c appropriately.
    * Makefile.in: Regenerated.
    * buildopscounttab.h: Fixes spelling mistake.
    * buildopscounttab.c: Fixes DW_OP_lit0 input.
      Adjusts indents, remove trailing white and trailing
      blank lines.
    * esb.c: Added ESBERR checks and that found mistakes
      in dwarfdump source.
    * print_die.c: include opscounttab.h. op_has_no_operands()
      is just  reference to a byte in dwarf_opscounttab array.
      Deleted a struct and table that is not used.
    * print_macro.c: Fixed a esb_append_printf_u() call
      missing a % .  Removed a stray newline from printf output.
    * print_ranges.c: An error message was missing
      a % to esb_append_printf_u. Fixed.
    * print_loclists.c: Rearranged 4 lines so it is easy
      to see that the printf has its 0x%  before DW_PR_XZEROS etc.

+++ scripts/ddbuild.sh
@@ -102,6 +102,23 @@ then
    echo tag_tree_build 4 compile fail
    exit 1
 fi
+
+$CC -I $top_srcdir/libdwarf \
+  $top_srcdir/dwarfdump/buildopscounttab.c \
+  $top_srcdir/dwarfdump/dwarf_names.c -o buildop
+if [ $? -ne 0 ]
+then
+    echo "FAIL compiling buildop  and building opstabcount.c source"
+    exit 1
+fi
+./buildop >opscounttab.c
+if [ $? -ne 0 ]
+then
+    echo "FAIL building opstabcount.c source"
+    exit 1
+fi
+rm -f ./buildop
+
 rm -f tmp-tag-tree-build4.tmp 
 rm -f tmp-t4.c
